+++ src/main/java/com/isa/morswiny/servlets/EventServlet.java
@@ -23,7 +23,7 @@ import java.util.Map;
 public class EventServlet extends HttpServlet {
 
     private static final Logger STDOUT = LoggerFactory.getLogger("CONSOLE_OUT");
-    private static final String TEMPLATE_NAME = "singleEvent.ftlh";
+    private static final String TEMPLATE_NAME = "singleEvent";
 
     @Inject
     EventCRUDRepositoryInterface eventCRUDRepositoryInterface;
