+++ src/main/webapp/WEB-INF/fm-templates/header.ftlh
@@ -7,7 +7,7 @@
 
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <div class="media">
-        <a href="main-page"><img src="/HTML/morswin.png" height="80" width="80" class="mr-3" alt="..."></a>
+        <a href="main-page"><img src="/Images/morswin.png" height="80" width="80" class="mr-3" alt="..."></a>
     </div>
     <a class="navbar-brand" href="#"><b>Morświny Events</b></a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
